ACK: Use varint
The interpretation of Additional ACK Blocks has to be changed in a
later commits.

+++ tests/ngtcp2_test_helper.h
@@ -66,14 +66,13 @@ size_t ngtcp2_t_encode_stream_frame(uint8_t *out, uint8_t flags,
 
 /*
  * ngtcp2_t_encode_ack_frame encodes ACK frame into |out| with the
- * given parameters.  Currently, this function encodes |largest_ack|
- * in 64 bits, and 1 ACK Block Section.  NumTS and ACK Delay fields
- * are always 0.
+ * given parameters.  Currently, this function encodes 1 ACK Block
+ * Section.  ACK Delay field is always 0.
  *
  * This function returns the number of bytes written to |out|.
  */
 size_t ngtcp2_t_encode_ack_frame(uint8_t *out, uint64_t largest_ack,
-                                 uint64_t first_ack_blklen, uint8_t gap,
+                                 uint64_t first_ack_blklen, uint64_t gap,
                                  uint64_t ack_blklen);
 
 /*
